+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added dependency to [SUOU](https://github.com/sakuragasaki46/suou) library
 - Added user blocks
 - Added user strikes: a strike logs the content of a removed message for future use
+- Posts may now be deleted by author. If it has comments, comments are not spared
 - Implemented guild subscriptions
     + Blocking a user prevents them from seeing your comments, posts (standalone or in feed) and profile
 - Added ✨color themes✨